Fixed build on OSX.
* Scene lab sample needs to build using C++11 for Flatbuffers.
* Fixed Scene lab sample build from the root of the project.
* Handled case where the assets directory is empty when building assets.

Tested:
Verified Scene Lab builds on OSX when generating an Xcode from the root
of the tree or under the samples directory.

Bug: 25866666

+++ scripts/scene_lab_asset_builder.py
@@ -1214,7 +1214,8 @@ def main(project_root='', assets_path='', asset_meta='', asset_roots=None,
   meta = parse_metadata(args.meta)
 
   if target != 'clean':
-    distutils.dir_util.copy_tree(assets_path, args.output, update=1)
+    if os.path.exists(assets_path):
+      distutils.dir_util.copy_tree(assets_path, args.output, update=1)
     for i in range(0, len(args.copy_tree), 2):
       distutils.dir_util.copy_tree(args.copy_tree[i], args.copy_tree[i + 1],
                                    update=1)
